Go back to using Reflections to fix some weird issues... also add det…

+++ src/main/java/gg/stephen/neptune/command/CommandDispatcher.java
@@ -4,48 +4,56 @@ import gg.stephen.neptune.annotation.Command;
 import gg.stephen.neptune.annotation.Inject;
 import gg.stephen.neptune.annotation.Instantiate;
 import gg.stephen.neptune.util.ArgumentConverter;
-import gg.stephen.neptune.util.ClassFinder;
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.entities.Guild;
+import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import net.dv8tion.jda.api.interactions.commands.DefaultMemberPermissions;
 import net.dv8tion.jda.api.interactions.commands.build.Commands;
 import net.dv8tion.jda.api.interactions.commands.build.SlashCommandData;
 import org.reflections.Reflections;
 import org.reflections.scanners.ResourcesScanner;
 import org.reflections.scanners.SubTypesScanner;
-import org.reflections.util.ClasspathHelper;
 import org.reflections.util.ConfigurationBuilder;
-import org.reflections.util.FilterBuilder;
 
 import java.io.IOException;
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.HashMap;
-import java.util.LinkedList;
-import java.util.List;
 import java.util.Set;
+import java.util.stream.Collectors;
+
+import static gg.stephen.neptune.Neptune.LOGGER;
 
 public class CommandDispatcher {
 
-    public CommandDispatcher(JDA jda, Object clazz, CommandManager manager, Guild... guilds) throws InstantiationException, IllegalAccessException, InvocationTargetException, IOException, ClassNotFoundException {
-        if (guilds != null) {
-            for (Guild guild : guilds) {
-                for (net.dv8tion.jda.api.interactions.commands.Command command : guild.retrieveCommands().complete()) {
-                    command.delete().complete();
+    public CommandDispatcher(JDA jda, Object clazz, CommandManager manager, boolean clearCommands, Guild... guilds) throws InstantiationException, IllegalAccessException, InvocationTargetException, IOException, ClassNotFoundException {
+        LOGGER.debug("Instantiating dispatcher...");
+        if (clearCommands) {
+            if (guilds != null) {
+                LOGGER.debug("Specified guilds, clearing all guild commands....");
+                int amount = 0;
+                for (Guild guild : guilds) {
+                    for (net.dv8tion.jda.api.interactions.commands.Command command : guild.retrieveCommands().complete()) {
+                        command.delete().complete();
+                        LOGGER.debug("Deleted " + command + " in " + guild + ".");
+                        amount++;
+                    }
                 }
-            }
-        } else {
-            for (net.dv8tion.jda.api.interactions.commands.Command command : jda.retrieveCommands().complete()) {
-                command.delete().complete();
+                LOGGER.debug("Cleared " + amount + " commands across " + guilds.length + " guilds.");
+            } else {
+                LOGGER.debug("No guilds were specified, clearing all commands globally...");
+                int amount = 0;
+                for (net.dv8tion.jda.api.interactions.commands.Command command : jda.retrieveCommands().complete()) {
+                    command.delete().complete();
+                    LOGGER.debug("Deleted " + command + " globally.");
+                    amount++;
+                }
+                LOGGER.debug("Cleared " + amount + " commands globally.");
             }
         }
 
         Class<?> classType = clazz.getClass();
-
-        List<ClassLoader> classLoadersList = new LinkedList<>();
-        classLoadersList.add(ClasspathHelper.contextClassLoader());
-        classLoadersList.add(ClasspathHelper.staticClassLoader());
 
         HashMap<String, Object> toInject = new HashMap<>();
         for (Method method : classType.getMethods()) {
@@ -54,21 +62,41 @@ public class CommandDispatcher {
             }
         }
 
-        Class[] classes = ClassFinder.getClasses(classType.getPackage().getName());
-        for (Class<?> foundClass : classes) {
+        LOGGER.debug("Found " + toInject.size() + " methods to inject with Neptune: " + toInject.keySet());
+
+        String packageName = classType.getPackage().getName();
+        LOGGER.debug("Scanning classes with related package \"" + packageName + "\".");
+
+        Reflections reflections = new Reflections(new ConfigurationBuilder()
+                .forPackages(packageName)
+                .setScanners(new SubTypesScanner(false), new ResourcesScanner())
+        );
+        Set<Class<?>> objects = reflections.getSubTypesOf(Object.class).stream()
+                .filter(aClass -> aClass.getPackage().getName().startsWith(packageName))
+                .collect(Collectors.toSet());
+        Set<Class<? extends ListenerAdapter>> listeners = reflections.getSubTypesOf(ListenerAdapter.class).stream()
+                .filter(aClass -> aClass.getPackage().getName().startsWith(packageName))
+                .collect(Collectors.toSet());
+
+        LOGGER.debug("Found " + objects.size() + " objects.");
+        LOGGER.debug("Found " + listeners.size() + " listeners.");
+
+        objects.addAll(listeners);
+
+        for (Class<?> foundClass : objects) {
             Object instance = foundClass.getName().equals(clazz.getClass().getName()) ? clazz : null;
             for (Method method : foundClass.getMethods()) {
                 if (method.isAnnotationPresent(Command.class)) {
-                    if (instance == null) {
-                        instance = foundClass.newInstance();
-                    }
+                    if (instance == null) instance = foundClass.newInstance();
+
                     Command command = method.getAnnotation(Command.class);
                     SlashCommandData commandData = Commands.slash(command.name(), command.description());
                     commandData.setDefaultPermissions(DefaultMemberPermissions.enabledFor(command.permissions()));
                     CommandMapping mapping = new CommandMapping(method, instance);
-                    for (CommandMapping.NamedParameter param : mapping.getParameters()) {
+
+                    for (CommandMapping.NamedParameter param : mapping.getParameters())
                         commandData.addOption(ArgumentConverter.toOptionType(param.getType()), param.getName(), param.getName(), param.isRequired());
-                    }
+
 
                     manager.addCommand(command.name(), mapping);
                     if (guilds != null) {
@@ -84,13 +112,14 @@ public class CommandDispatcher {
             for (Field field : foundClass.getDeclaredFields()) {
                 if (instance == null && field.isAnnotationPresent(Inject.class)) {
                     instance = foundClass.newInstance();
+                    LOGGER.debug("Instantiated " + instance.getClass().getSimpleName() + " due to it's @Inject annotation usage.");
                 }
             }
 
             if (instance != null) {
                 for (Field field : instance.getClass().getDeclaredFields()) {
                     if (field.isAnnotationPresent(Inject.class)) {
-                        System.out.println("injected " + field.getName() + " into " + instance.getClass().getSimpleName());
+                        LOGGER.debug("Injected " + field.getName() + " into " + instance.getClass().getSimpleName() + ".");
                         field.setAccessible(true);
                         field.set(instance, toInject.get(field.getName()));
                     }
@@ -98,7 +127,7 @@ public class CommandDispatcher {
 
                 if (instance.getClass().getSuperclass().getName().equals("net.dv8tion.jda.api.hooks.ListenerAdapter")) {
                     jda.addEventListener(instance);
-                    System.out.println("added " + instance.getClass().getSimpleName() + " as an EventListener");
+                    LOGGER.debug("Automatically registered " + instance.getClass().getSimpleName() + " as an EventListener due to it having an injected dependency.");
                 }
             }
         }
